+++ client/src/main/resources/main.fxml
@@ -9,12 +9,12 @@
 <?import javafx.scene.control.TextField?>
 <VBox alignment="CENTER" prefHeight="400.0" prefWidth="400.0"
       xmlns:fx="http://javafx.com/fxml/1" xmlns="http://javafx.com/javafx/8.0.172-ea" fx:controller="ClientController">
-    <VBox fx:id="authPanel" alignment="CENTER" HBox.hgrow="ALWAYS" VBox.vgrow="ALWAYS" visible="true">
-        <Label fx:id="authMessage" visible="false"></Label>
+    <Label fx:id="authMessage" visible="false"></Label>
+    <HBox fx:id="authPanel" HBox.hgrow="ALWAYS" VBox.vgrow="ALWAYS" visible="true">
         <TextField fx:id="tfUsername" promptText="Имя пользователя" />
         <TextField fx:id="tfPassword" promptText="Пароль" />
         <Button onAction="#pressOnAuthBtn" text="Авторизоваться" HBox.hgrow="ALWAYS" />
-    </VBox>
+    </HBox>
     <HBox fx:id="clientPanel" HBox.hgrow="ALWAYS" VBox.vgrow="ALWAYS" visible="false">
         <VBox fx:id="vbClient" HBox.hgrow="ALWAYS">
             <Label>Client</Label>
