+++ server/src/main/java/Server.java
@@ -18,7 +18,6 @@ public class Server {
                     .childHandler(new ChannelInitializer<SocketChannel>() {
                         protected void initChannel(SocketChannel socketChannel) throws Exception {
                             socketChannel.pipeline().addLast(
-                                    new AuthHandler(),
                                     new MainHandler()
                             );
                         }
